[releng] Move to Sirius 7.0.0-S20210908-102403
- Adapt to Sirius Modularization Changes
- Use the same url than Sirius for the Orbit location to avoid conflicts
in add-ons target platforms.

Bug: http://eclipse.se/531282
