+++ core/plugins/org.polarsys.capella.core.platform.sirius.clipboard/src/org/polarsys/capella/core/platform/sirius/clipboard/util/SiriusUtil.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2006, 2020 THALES GLOBAL SERVICES.
+ * Copyright (c) 2006, 2021 THALES GLOBAL SERVICES.
  * 
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -25,7 +25,6 @@ import org.eclipse.sirius.diagram.DDiagram;
 import org.eclipse.sirius.diagram.DDiagramElement;
 import org.eclipse.sirius.diagram.DNode;
 import org.eclipse.sirius.diagram.DNodeContainer;
-import org.eclipse.sirius.diagram.business.internal.metamodel.helper.LayerHelper;
 import org.eclipse.sirius.diagram.description.AbstractNodeMapping;
 import org.eclipse.sirius.diagram.description.ContainerMapping;
 import org.eclipse.sirius.diagram.description.ContainerMappingImport;
@@ -35,6 +34,7 @@ import org.eclipse.sirius.diagram.description.DiagramElementMapping;
 import org.eclipse.sirius.diagram.description.Layer;
 import org.eclipse.sirius.diagram.description.NodeMapping;
 import org.eclipse.sirius.diagram.description.NodeMappingImport;
+import org.eclipse.sirius.diagram.model.business.internal.helper.LayerModelHelper;
 import org.eclipse.sirius.diagram.sequence.SequenceDDiagram;
 import org.eclipse.sirius.viewpoint.DRefreshable;
 import org.eclipse.sirius.viewpoint.DSemanticDecorator;
@@ -155,7 +155,7 @@ public final class SiriusUtil {
     }
     // For all layers of the DiagramDescription, get all
     // AbstractNodeMappings which deal with the root of diagrams
-    EList<Layer> layers = LayerHelper.getAllLayers(desc);
+    EList<Layer> layers = LayerModelHelper.getAllLayers(desc);
     for (Layer layer : layers) {
       result.addAll(layer.getContainerMappings());
       result.addAll(layer.getNodeMappings());
