+++ core/plugins/org.polarsys.capella.core.sirius.analysis/src/org/polarsys/capella/core/sirius/analysis/refresh/extension/ContextualCapabilityRefreshExtension.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2006, 2020 THALES GLOBAL SERVICES.
+ * Copyright (c) 2006, 2021 THALES GLOBAL SERVICES.
  * 
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -20,11 +20,10 @@ import java.util.Set;
 import org.eclipse.sirius.common.tools.api.interpreter.IInterpreter;
 import org.eclipse.sirius.diagram.DDiagram;
 import org.eclipse.sirius.diagram.DDiagramElement;
-import org.eclipse.sirius.diagram.business.internal.metamodel.description.extensions.INodeMappingExt;
 import org.eclipse.sirius.diagram.business.internal.metamodel.helper.NodeMappingHelper;
 import org.eclipse.sirius.diagram.description.NodeMapping;
+import org.eclipse.sirius.tools.api.SiriusPlugin;
 import org.eclipse.sirius.viewpoint.DSemanticDecorator;
-import org.eclipse.sirius.viewpoint.SiriusPlugin;
 import org.polarsys.capella.core.data.capellacore.GeneralizableElement;
 import org.polarsys.capella.core.data.cs.Component;
 import org.polarsys.capella.core.data.ctx.Capability;
@@ -88,7 +87,7 @@ public class ContextualCapabilityRefreshExtension extends AbstractCacheAwareRefr
     for (final Component actor : actors) {
 
       if (!DiagramServices.getDiagramServices().isOnDiagram(diagram, actor)) {
-        final DDiagramElement container = nodeMappingHelper.createNode((INodeMappingExt) actorNodeMapping, actor, capa,
+        final DDiagramElement container = nodeMappingHelper.createNode(actorNodeMapping, actor, capa,
             diagram);
         diagram.getOwnedDiagramElements().add(container);
       }
@@ -96,7 +95,7 @@ public class ContextualCapabilityRefreshExtension extends AbstractCacheAwareRefr
 
     for (final AbstractCapability cap : capabilities) {
       if (!DiagramServices.getDiagramServices().isOnDiagram(diagram, cap)) {
-        final DDiagramElement container = nodeMappingHelper.createNode((INodeMappingExt) capaNodeMapping, cap, capa,
+        final DDiagramElement container = nodeMappingHelper.createNode(capaNodeMapping, cap, capa,
             diagram);
         diagram.getOwnedDiagramElements().add(container);
       }
@@ -104,7 +103,7 @@ public class ContextualCapabilityRefreshExtension extends AbstractCacheAwareRefr
 
     for (final Mission mission : missions) {
       if (!DiagramServices.getDiagramServices().isOnDiagram(diagram, mission)) {
-        final DDiagramElement container = nodeMappingHelper.createNode((INodeMappingExt) missionNodeMapping, mission,
+        final DDiagramElement container = nodeMappingHelper.createNode(missionNodeMapping, mission,
             capa, diagram);
         diagram.getOwnedDiagramElements().add(container);
       }
