+++ releng/plugins/org.polarsys.capella.targets/full/capella.target-definition.targetplatform
@@ -12,7 +12,7 @@
  *******************************************************************************/
 target "Capella"
 
-include "https://download.eclipse.org/sirius/updates/stable/7.0.0-S20210722-123507/2021-06/targets/modules/gmf-runtime-1.14.tpd"
+include "https://download.eclipse.org/sirius/updates/stable/7.0.0-S20210908-102403/2021-06/targets/modules/gmf-runtime-1.14.tpd"
 
 with source, requirements
 
@@ -43,7 +43,7 @@ location eclipse "http://download.eclipse.org/releases/2021-06" {
 	com.google.guava
 }
 
-location sirius "https://download.eclipse.org/sirius/updates/stable/7.0.0-S20210722-123507/2021-06" {
+location sirius "https://download.eclipse.org/sirius/updates/stable/7.0.0-S20210908-102403/2021-06" {
 	org.eclipse.sirius.doc.feature.feature.group
 	org.eclipse.sirius.doc.feature.source.feature.group
 	org.eclipse.sirius.runtime.source.feature.group
@@ -56,7 +56,7 @@ location sirius "https://download.eclipse.org/sirius/updates/stable/7.0.0-S20210
 	org.eclipse.sirius.runtime.ide.ui.acceleo.source.feature.group
 }
 
-location Orbit "https://download.eclipse.org/tools/orbit/downloads/2021-06" {
+location Orbit "https://download.eclipse.org/tools/orbit/downloads/drops/R20210602031627/repository" {
 	org.apache.batik.bridge.source lazy
 	org.apache.batik.script.source lazy
 }
