+++ tests/plugins/org.polarsys.capella.test.odesign/src/org/polarsys/capella/test/odesign/ju/titleblock/CheckTitleBlockMappingTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2020 THALES GLOBAL SERVICES.
+ * Copyright (c) 2020, 2021 THALES GLOBAL SERVICES.
  * 
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -13,11 +13,9 @@
 package org.polarsys.capella.test.odesign.ju.titleblock;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.stream.Collectors;
 
 import org.eclipse.sirius.business.api.componentization.ViewpointRegistry;
-import org.eclipse.sirius.diagram.business.internal.metamodel.helper.ContentHelper;
 import org.eclipse.sirius.diagram.description.AbstractNodeMapping;
 import org.eclipse.sirius.diagram.description.DiagramDescription;
 import org.eclipse.sirius.diagram.description.DiagramElementMapping;
@@ -27,7 +25,6 @@ import org.polarsys.capella.core.sirius.analysis.DiagramServices;
 import org.polarsys.capella.core.sirius.analysis.IDiagramNameConstants;
 import org.polarsys.capella.core.sirius.analysis.IMappingNameConstants;
 import org.polarsys.capella.core.sirius.analysis.IViewpointNameConstants;
-import org.polarsys.capella.test.framework.api.BasicTestCase;
 import org.polarsys.capella.test.odesign.helper.OdesignTestHelper;
 
 public class CheckTitleBlockMappingTest extends CheckTitleBlockTest {
